requirements.txt
asyncpg==0.29.0
pandas==2.2.0
matplotlib==3.8.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
geoip2==4.8.0
python-dotenv==1.0.1
jupyter==1.0.0
ipykernel==6.29.0